refactor: Extract GCP machine extractor and resource price loader
